+++ src/domain/core/Cargo.toml
@@ -34,7 +34,6 @@ container-runtime = { workspace = true }
 file-utils = { workspace = true }
 internal-error = { workspace = true }
 kamu-datasets = { workspace = true }
-messaging-outbox = { workspace = true }
 odf = { workspace = true }
 
 async-stream = { version = "0.3", default-features = false }
