--- src/domain/core/src/messages/core_message_consumers.rs
@@ -1,15 +1,0 @@
-// Copyright Kamu Data, Inc. and contributors. All rights reserved.
-//
-// Use of this software is governed by the Business Source License
-// included in the LICENSE file.
-//
-// As of the Change Date specified in that file, in accordance with
-// the Business Source License, use of this software will be governed
-// by the Apache License, Version 2.0.
-
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-pub const MESSAGE_CONSUMER_KAMU_CORE_DATASET_OWNERSHIP_SERVICE: &str =
-    "dev.kamu.domain.core.services.DatasetOwnershipService";
-
-////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
