+++ src/domain/datasets/domain/src/lib.rs
@@ -12,11 +12,13 @@
 #![feature(let_chains)]
 
 mod entities;
+mod messages;
 mod repos;
 mod services;
 mod use_cases;
 
 pub use entities::*;
+pub use messages::*;
 pub use repos::*;
 pub use services::*;
 pub use use_cases::*;
