+++ src/domain/datasets/domain/src/messages/datasets_message_producers.rs
@@ -9,7 +9,6 @@
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-pub const MESSAGE_PRODUCER_KAMU_CORE_DATASET_SERVICE: &str =
-    "dev.kamu.domain.core.services.DatasetService";
+pub const MESSAGE_PRODUCER_KAMU_DATASET_SERVICE: &str = "dev.kamu.domain.datasets.DatasetService";
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
