+++ src/domain/flow-system/services/tests/tests/test_flow_configuration_service_impl.rs
@@ -16,7 +16,12 @@ use futures::TryStreamExt;
 use kamu::*;
 use kamu_accounts::CurrentAccountSubject;
 use kamu_core::*;
-use kamu_datasets::{CreateDatasetFromSnapshotUseCase, DeleteDatasetUseCase};
+use kamu_datasets::{
+    CreateDatasetFromSnapshotUseCase,
+    DatasetLifecycleMessage,
+    DeleteDatasetUseCase,
+    MESSAGE_PRODUCER_KAMU_DATASET_SERVICE,
+};
 use kamu_datasets_inmem::InMemoryDatasetDependencyRepository;
 use kamu_datasets_services::{
     CreateDatasetFromSnapshotUseCaseImpl,
@@ -242,7 +247,7 @@ impl FlowConfigurationHarness {
 
             register_message_dispatcher::<DatasetLifecycleMessage>(
                 &mut b,
-                MESSAGE_PRODUCER_KAMU_CORE_DATASET_SERVICE,
+                MESSAGE_PRODUCER_KAMU_DATASET_SERVICE,
             );
             register_message_dispatcher::<FlowConfigurationUpdatedMessage>(
                 &mut b,
