+++ src/infra/core/src/testing/outbox_expectation_helpers.rs
@@ -7,7 +7,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use kamu_core::{DatasetLifecycleMessage, MESSAGE_PRODUCER_KAMU_CORE_DATASET_SERVICE};
+use kamu_datasets::{DatasetLifecycleMessage, MESSAGE_PRODUCER_KAMU_DATASET_SERVICE};
 use messaging_outbox::MockOutbox;
 use mockall::predicate::{always, eq, function};
 
@@ -17,7 +17,7 @@ pub fn expect_outbox_dataset_created(mock_outbox: &mut MockOutbox, times: usize)
     mock_outbox
         .expect_post_message_as_json()
         .with(
-            eq(MESSAGE_PRODUCER_KAMU_CORE_DATASET_SERVICE),
+            eq(MESSAGE_PRODUCER_KAMU_DATASET_SERVICE),
             function(|message_as_json: &serde_json::Value| {
                 matches!(
                     serde_json::from_value::<DatasetLifecycleMessage>(message_as_json.clone()),
@@ -36,7 +36,7 @@ pub fn expect_outbox_dataset_dependencies_updated(mock_outbox: &mut MockOutbox, 
     mock_outbox
         .expect_post_message_as_json()
         .with(
-            eq(MESSAGE_PRODUCER_KAMU_CORE_DATASET_SERVICE),
+            eq(MESSAGE_PRODUCER_KAMU_DATASET_SERVICE),
             function(|message_as_json: &serde_json::Value| {
                 matches!(
                     serde_json::from_value::<DatasetLifecycleMessage>(message_as_json.clone()),
@@ -55,7 +55,7 @@ pub fn expect_outbox_dataset_renamed(mock_outbox: &mut MockOutbox, times: usize)
     mock_outbox
         .expect_post_message_as_json()
         .with(
-            eq(MESSAGE_PRODUCER_KAMU_CORE_DATASET_SERVICE),
+            eq(MESSAGE_PRODUCER_KAMU_DATASET_SERVICE),
             function(|message_as_json: &serde_json::Value| {
                 matches!(
                     serde_json::from_value::<DatasetLifecycleMessage>(message_as_json.clone()),
@@ -74,7 +74,7 @@ pub fn expect_outbox_dataset_deleted(mock_outbox: &mut MockOutbox, times: usize)
     mock_outbox
         .expect_post_message_as_json()
         .with(
-            eq(MESSAGE_PRODUCER_KAMU_CORE_DATASET_SERVICE),
+            eq(MESSAGE_PRODUCER_KAMU_DATASET_SERVICE),
             function(|message_as_json: &serde_json::Value| {
                 matches!(
                     serde_json::from_value::<DatasetLifecycleMessage>(message_as_json.clone()),
